refactor: extract context from Main
순환 참조를 없애기 위해 메인페이지에서 컨텍스트 추출
그 후 경로 수정
