+++ front/src/Router.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import Home from './pages/Home';
-import Main from './pages/Main';
+import Main from './pages/Main/Main';
 import MyPage from './pages/MyPage';
 
 const Router = () => (
