+++ front/src/pages/Main/Main.context.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const MainSocketContext = React.createContext();
+
+export default MainSocketContext;
